+++ java/integTest/src/test/java/glide/SharedCommandTests.java
@@ -4376,8 +4376,8 @@ public class SharedCommandTests {
     @MethodSource("getClients")
     public void lset(BaseClient client) {
         // setup
-        String key = "testKey";
-        String nonExistingKey = "nonExisting";
+        String key = UUID.randomUUID().toString();
+        String nonExistingKey = UUID.randomUUID().toString();
         long index = 0;
         long oobIndex = 10;
         long negativeIndex = -1;
@@ -4949,5 +4949,36 @@ public class SharedCommandTests {
         ExecutionException executionException =
                 assertThrows(ExecutionException.class, () -> client.sintercard(badArr).get());
         assertInstanceOf(RequestException.class, executionException.getCause());
+    }
+
+    @SneakyThrows
+    @ParameterizedTest(autoCloseArguments = false)
+    @MethodSource("getClients")
+    public void copy(BaseClient client) {
+        assumeTrue(REDIS_VERSION.isGreaterThanOrEqualTo("6.2.0"), "This feature added in redis 6.2.0");
+        // setup
+        String source = "{key}-1" + UUID.randomUUID();
+        String destination = "{key}-2" + UUID.randomUUID();
+
+        // neither key exists, returns false
+        assertFalse(client.copy(source, destination, false).get());
+        assertFalse(client.copy(source, destination).get());
+
+        // source exists, destination does not
+        client.set(source, "one");
+        assertTrue(client.copy(source, destination, false).get());
+        assertEquals("one", client.get(destination).get());
+
+        // setting new value for source
+        client.set(source, "two");
+
+        // both exists, no REPLACE
+        assertFalse(client.copy(source, destination).get());
+        assertFalse(client.copy(source, destination, false).get());
+        assertEquals("one", client.get(destination).get());
+
+        // both exists, with REPLACE
+        assertTrue(client.copy(source, destination, true).get());
+        assertEquals("two", client.get(destination).get());
     }
 }